✨ (Message): refactor message components to improve structure and add Reactions component for better reaction handling 📝 (ItemViewer): add ItemViewer component and its Storybook configuration for better visualization and testing of message interactions ♻️ (MessageText): clean up unused code and improve imports for better maintainability
