+++ src/Molecules/Message/MessageItem/components/MessageText/MessageText.tsx
@@ -1,18 +1,13 @@
 import { Box, Typography, styled } from "@mui/material";
-import type { MessageType } from "../../../types";
+import type { MessageType, Reaction } from "../../../types";
 import {
   bubbleBackgroundColors,
   messageBubbleBorderBottomLeftRadius,
   messageBubbleBorderBottomRightRadius,
   messageContainerWidth,
   messageTextColors,
-  reactionsContainerJustifyContent,
 } from "./styles/maps";
-
-export interface Reaction {
-  icon: string;
-  count: number;
-}
+import { Reactions } from "../Reactions/Reactions";
 
 interface TextProps {
   children: React.ReactNode;
@@ -45,55 +40,6 @@ const MessageContent = styled(Typography, {
   fontSize: "16px",
 }));
 
-const ReactionsContainer = styled(Box, {
-  shouldForwardProp: (prop) => prop !== "messageType",
-})<{ messageType: MessageType }>(({ messageType }) => ({
-  display: "flex",
-  justifyContent: reactionsContainerJustifyContent[messageType],
-}));
-
-const ReactionsWrapper = styled(Box)({
-  marginTop: "3px",
-  width: "fit-content",
-  padding: "2px 8px",
-  borderRadius: "100px",
-  backgroundColor: "#EFF0F2",
-  display: "flex",
-  alignItems: "center",
-  gap: "4px",
-});
-
-const ReactionItem = styled(Box)({
-  display: "flex",
-  alignItems: "center",
-  gap: "3.2px",
-});
-
-interface ReactionsProps {
-  reactions: Reaction[];
-  messageType: MessageType;
-}
-
-const Reactions = ({ reactions, messageType }: ReactionsProps) => {
-  if (reactions.length === 0) {
-    return null;
-  }
-
-  return (
-    <ReactionsContainer messageType={messageType}>
-      <ReactionsWrapper>
-        {reactions.map((reaction) => (
-          <ReactionItem key={reaction.icon}>
-            <Box>{reaction.icon}</Box>
-            <Typography variant="caption" color="black">
-              {reaction.count}
-            </Typography>
-          </ReactionItem>
-        ))}
-      </ReactionsWrapper>
-    </ReactionsContainer>
-  );
-};
 
 export const MessageText = ({
   children,
