+++ src/Molecules/Message/MessageItem/components/index.ts
@@ -1,1 +1,2 @@
 export * from "./MessageText/MessageText";
+export * from "./Reactions/Reactions";
